requirements.txt
uvicorn~=0.41.0
fastapi~=0.133.1
langchain-core~=1.2.14
langchain-openai~=1.1.10
langserve~=0.3.3
python-dotenv~=1.2.1
requests~=2.32.5
streamlit~=1.54.0